VCFs: Comparing bcftools filtering
This is a push of my python functions and shells for filtering the indels out of vcf chromosomes with different bcftools methods, then comparing the results in various ways. Also includes a README in markdown that explains what each function does. Also includes folders with the .txt file results of the filtering and comparisons, as well as some sanity-check testing files.